# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Epplaa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Wallet" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/topup:
    post:
      operationId: walletTopUp
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountMinor]
              properties:
                amountMinor: { type: integer }
                label: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WalletTopupResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { description: Top-up payment init failed }
  /wallet/spend:
    post:
      operationId: walletSpend
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountMinor, label]
              properties:
                amountMinor: { type: integer }
                label: { type: string }
                refId: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Wallet" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/withdraw:
    post:
      operationId: walletWithdraw
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountMinor, destinationLabel]
              properties:
                amountMinor: { type: integer }
                destinationLabel: { type: string }
                bankCode: { type: string }
                bankLast4: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WalletWithdrawResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/promo-credit:
    post:
      operationId: walletPromoCredit
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountMinor, label]
              properties:
                amountMinor: { type: integer }
                label: { type: string }
                refId: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Wallet" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/refund:
    post:
      operationId: walletRefund
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [returnId, amountMinor, label]
              properties:
                returnId: { type: string }
                amountMinor: { type: integer }
                label: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Wallet" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /wallet/settings:
    patch:
      operationId: updateWalletSettings
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currencyCode]
              properties:
                currencyCode: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Wallet" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  responses:
    Unauthorized:
      description: Sign-in required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: string }
...
    WalletTxn:
      type: object
      required: [id, kind, amountMinor, label, atIso]
      properties:
        id: { type: string }
        kind: { type: string }
        amountMinor: { type: integer }
        label: { type: string }
        refId: { type: ["string", "null"] }
        atIso: { type: string }
    Wallet:
      type: object
      required: [balanceMinor, currencyCode, txns]
      properties:
        balanceMinor: { type: integer }
        currencyCode: { type: string }
        txns:
          type: array
          items: { $ref: "#/components/schemas/WalletTxn" }
    WalletTopupResponse:
      type: object
      required: [balanceMinor, currencyCode, txns]
      properties:
        balanceMinor: { type: integer }
        currencyCode: { type: string }
        txns: { type: array, items: { $ref: "#/components/schemas/WalletTxn" } }
        pendingTopup:
          type: object
          required: [intentId, reference, gateway, amountMinor, status]
          properties:
            intentId: { type: string }
            reference: { type: string }
            gateway: { type: string }
            amountMinor: { type: integer }
            authorizationUrl: { type: ["string", "null"] }
            status: { type: string }
    WalletWithdrawResponse:
      type: object
      required: [balanceMinor, currencyCode, txns]
      properties:
        balanceMinor: { type: integer }
        currencyCode: { type: string }
        txns: { type: array, items: { $ref: "#/components/schemas/WalletTxn" } }
        pendingWithdrawal:
          type: object
          required: [payoutId, reference, amountMinor, status]
          properties:
            payoutId: { type: string }
            reference: { type: string }
            amountMinor: { type: integer }
            status: { type: string }
...